rename upsert_document to upsert_queue
